requirements.txt
py-cord
keyboard
python-dotenv
pyautogui